move to entity_recognizer.py
