+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/es_nif_recognizer.py
@@ -1,7 +1,6 @@
 from typing import List, Optional, Tuple
 
-from presidio_analyzer import Pattern, PatternRecognizer
-from presidio_analyzer.validation import ValidationUtils
+from presidio_analyzer import EntityRecognizer, Pattern, PatternRecognizer
 
 
 class EsNifRecognizer(PatternRecognizer):
@@ -48,7 +47,7 @@ class EsNifRecognizer(PatternRecognizer):
         )
 
     def validate_result(self, pattern_text: str) -> bool:  # noqa D102
-        pattern_text = ValidationUtils.sanitize_value(
+        pattern_text = EntityRecognizer.sanitize_value(
             pattern_text, self.replacement_pairs
         )
         letter = pattern_text[-1]
