+++ presidio-analyzer/presidio_analyzer/predefined_recognizers/in_vehicle_registration_recognizer.py
@@ -1,7 +1,6 @@
 from typing import List, Optional, Tuple
 
-from presidio_analyzer import Pattern, PatternRecognizer
-from presidio_analyzer.validation import ValidationUtils
+from presidio_analyzer import EntityRecognizer, Pattern, PatternRecognizer
 
 
 class InVehicleRegistrationRecognizer(PatternRecognizer):
@@ -349,7 +348,7 @@ class InVehicleRegistrationRecognizer(PatternRecognizer):
 
     def validate_result(self, pattern_text: str) -> bool:
         """Determine absolute value based on calculation."""
-        sanitized_value = ValidationUtils.sanitize_value(
+        sanitized_value = EntityRecognizer.sanitize_value(
             pattern_text, self.replacement_pairs
         )
         return self.__check_vehicle_registration(sanitized_value)
