--- presidio-analyzer/presidio_analyzer/validation/__init__.py
@@ -1,5 +1,0 @@
-"""Validation."""
-
-from .validation_utils import ValidationUtils
-
-__all__ = ["ValidationUtils"]
