+++ presidio-analyzer/tests/test_entity_recognizer.py
@@ -119,3 +119,23 @@ def test_when_remove_duplicates_contained_shorter_length_results_removed():
     ]
     results = EntityRecognizer.remove_duplicates(arr)
     assert len(results) == 1
+
+from presidio_analyzer.validation import ValidationUtils
+import pytest
+
+sanitizer_test_set = [
+    ["  a|b:c       ::-", [("-", ""), (" ", ""), (":", ""), ("|", "")], "abc"],
+    ["def", "", "def"],
+]
+
+@pytest.mark.parametrize("input_text, params, expected_output", sanitizer_test_set)
+def test_sanitize_value(input_text, params, expected_output):
+    """
+    Test to assert sanitize_value functionality from base class.
+
+    :param input_text: input string
+    :param params: List of tuples, indicating what has to be sanitized with which
+    :param expected_output: sanitized value
+    :return: True/False
+    """
+    assert ValidationUtils.sanitize_value(input_text, params) == expected_output
