--- presidio-analyzer/tests/test_validation_utils.py
@@ -1,19 +1,0 @@
-from presidio_analyzer.validation import ValidationUtils
-import pytest
-
-sanitizer_test_set = [
-    ["  a|b:c       ::-", [("-", ""), (" ", ""), (":", ""), ("|", "")], "abc"],
-    ["def", "", "def"],
-]
-
-@pytest.mark.parametrize("input_text, params, expected_output", sanitizer_test_set)
-def test_sanitize_value(input_text, params, expected_output):
-    """
-    Test to assert sanitize_value functionality from base class.
-
-    :param input_text: input string
-    :param params: List of tuples, indicating what has to be sanitized with which
-    :param expected_output: sanitized value
-    :return: True/False
-    """
-    assert ValidationUtils.sanitize_value(input_text, params) == expected_output
